Add files via upload
This function file2, takes input from file.dat and outputs it to the screen.   I believe I am using class and encapsulation.  It is an exercise . I would also like to learn more about function output and arguments as well as pointer and memory allocation.

I am putting this up here hoping someone may help,  it is like a message in a bottle.